pattern_dump: Add a [%hdlcd] format for dumping HDLC payload data bytes
Add a [%hdlcd] format string for dumping HDLC payload data bytes
in hex.  The bytes are un-zero-bit-stuffed and bit reversed, so the
payload can be read easily.

This format is useful if you have an AIS or AX.25 packet that has
bit errors, but you want to partially decode and inspect as much as you
can anyway.

TODO: This change doesn't bother to sensibly handle more than 5 ones
in a row, including the HDLC flag character (0x7e).